+++ next/listtransactions/doc.go
@@ -1,0 +1,3 @@
+// Package listtransactions shows all transactions
+
+package listtransactions
