+++ next/listtransactions/errors.go
@@ -1,4 +1,4 @@
-package report
+package listtransactions
 
 import "fmt"
