+++ next/listtransactions/interfaces.go
@@ -1,4 +1,4 @@
-package report
+package listtransactions
 
 import "github.com/luistm/banksaurus/next/entity/transaction"
 
@@ -7,7 +7,7 @@ type Presenter interface {
 	Present([]map[string]int64) error
 }
 
-// Repository is a collection of entities to be used by the report
+// Repository is a collection of entities to be used by the listtransactions
 type Repository interface {
 	GetAll() ([]*transaction.Entity, error)
 }
