--- next/report/doc.go
@@ -1,3 +1,0 @@
-// Package report shows all transactions
-
-package report
